Теперь можно видеть свою статистику правильнх ответов за последние пять попыток в каждом wordlist'е

+++ KanjiReaderBack/src/main/scala/kanjiReader/leveling/LevelDomain.scala
@@ -42,7 +42,7 @@ object QuestType extends Enumeration {
 }
 
 case class WordGameResult(
-    wordList: Int,
+    wordList: Byte,
     time: Int,
     count: Int,
     correctCount: Int,
